+++ esmerald/routing/views.py
@@ -1,4 +1,0 @@
-from .apis import APIView, SimpleAPIView
-from .apis.base import View
-
-__all__ = ["APIView", "SimpleAPIView", "View"]